pyproject.toml
[tool.poetry]
name = "quiz-system"
version = "0.1.0"
description = "A Flask-based quiz system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.0.0"
flask-sqlalchemy = "^3.0.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
werkzeug = "^2.0.0"
PyPDF2 = "^3.0.0"
python-docx = "^0.8.11"
pymysql = "^1.0.0"
aiohttp = "^3.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
flake8 = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
